chore: move ValidatePluginCoreCompatibility into ProcessPluginLoadContext
